requirements.txt
fastapi
pydantic
pyjwt[crypto]
pyodbc
python-multipart
requests
starlette
sqlalchemy
